refactor(modes): move start and end helpers to src/
